Add helper script for import manipulation
When developing, one often wishes to check if new changes broke any of
the old blueprints. In order to make this a bit easier to do, we added
a simple script that does the redirection.

This script can also be used when preparing release, since script does
not care if redirection is done to a branch or tag.

Change-Id: I66d6fa7d30d94e822b8b0995134a1e2334d25b30