Add simple data pipeline implementation
This blueprint exploits the generic shell runner node to run use case
provider's custom shell execution. At the same time it contains all
the needed wiring to ensure that Cloudify's orchestration work will
produce parameters needed by the shell scripts.

In order to provide as authentic deployment as possible, we created a
dummy services that serve as a (fully functional) placeholders. This
made it possible to create an application that actually produces some
data and passes it around.

The accompanying documentation should make it relatively clear why
things are done in a certain way and how various pieces fit together.

Change-Id: Ibbcf8d4429eb07675f34de5e7d62e69c9e8224c2