Fix sharded MongoDB blueprint
When we wrote the blueprint, we misplaced the bind_ip property and set
it on host type instead of the router, which caused validation to
fail. This commit makes blueprint pass the validation.

Note that blueprint will still fail to deploy properly, since MongoDB
cluster is not fully supported yet.

Change-Id: I4f0d6aceb42ff17b0a243d5331f97bf5646fc585